Restructure as package of tools with pyproject
