+++ src/kabr_tools/__init__.py
@@ -1,0 +1,3 @@
+from kabr_tools import cvat2slowfast, cvat2ultralytics, detector2cvat, player, tracks_extractor
+
+__all__ = ["cvat2slowfast", "cvat2ultralytics", "detector2cvat", "player", "tracks_extractor"]
