+++ src/kabr_tools/detector2cvat.py
@@ -2,10 +2,11 @@ import os
 import argparse
 import cv2
 from tqdm import tqdm
-from src.yolo import YOLOv8
-from src.tracker import Tracker, Tracks
-from src.object import Object
-from src.draw import Draw
+from kabr_tools.utils.yolo import YOLOv8
+from kabr_tools.utils.tracker import Tracker, Tracks
+from kabr_tools.utils.object import Object
+from kabr_tools.utils.draw import Draw
+
 
 
 def detector2cvat(path_to_videos, path_to_save):
@@ -113,6 +114,10 @@ def parse_args():
     return local_parser.parse_args()
 
 
-if __name__ == "__main__":
+def main():
     args = parse_args()
     detector2cvat(args.video, args.save)
+
+
+if __name__ == "__main__":
+    main()
